Remove location linking parts from osm import.
